Modify backend to add timestamps proper formatting, and fix intake form validation error:
The dynamically rendered stake holders for the intake form were not
having their error message show up correct, I fix an issue in the
validation method which resolved that issue.

I also added a timestamp for each submission into the submission log,
this will make it so if there is ever a need to search the log, we can
do so without much issue.